AiSopp/AiSopp/requirements.txt
Flask<1
numpy
tensorflow
maxminddb-geolite2
flask_uploads
flask-wtf
